+++ game/tests/test_views_common.py
@@ -10,6 +10,7 @@ from django.urls import reverse
 import game.forms as gforms
 import game.models as gmodels
 import game.views.common as gvcommon
+import master.models as mmodels
 from game.tests import utils
 
 
@@ -389,40 +390,36 @@ class CreateGameViewTest(TestCase):
         user = User.objects.create(username=utils.generate_random_name(5))
         user.set_password("pwd")
         user.save()
+        mmodels.Story.objects.create(title=utils.generate_random_name(20))
 
     def setUp(self):
         self.user = User.objects.last()
         self.client.login(username=self.user.username, password="pwd")
+        self.story = mmodels.Story.objects.last()
 
     def test_view_mapping(self):
-        response = self.client.get(reverse(self.path_name))
+        response = self.client.get(reverse(self.path_name, args=(self.story.slug,)))
         self.assertEqual(response.status_code, 200)
         self.assertEqual(
             response.resolver_match.func.view_class, gvcommon.CreateGameView
         )
 
     def test_template_mapping(self):
-        response = self.client.get(reverse(self.path_name))
+        response = self.client.get(reverse(self.path_name, args=(self.story.slug,)))
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, "game/creategame.html")
 
     def test_game_creation(self):
-        name = utils.generate_random_name(20)
-        description = utils.generate_random_string(100)
-        data = {"name": f"{name}", "description": f"{description}"}
-        form = gforms.CreateGameForm(data)
-        self.assertTrue(form.is_valid())
-
-        response = self.client.post(reverse(self.path_name), data=form.cleaned_data)
+        response = self.client.post(reverse(self.path_name, args=(self.story.slug,)))
         self.assertEqual(response.status_code, 302)
         game = gmodels.Game.objects.last()
-        self.assertEqual(game.name, name)
+        self.assertEqual(game.name, self.story.title)
         self.assertEqual(game.status, "P")
         self.assertEqual(game.master, self.user)
         tale = gmodels.Tale.objects.last()
         self.assertEqual(tale.game, game)
         self.assertEqual(tale.message, "The Master created the story.")
-        self.assertEqual(tale.content, form.cleaned_data["description"])
+        self.assertEqual(tale.content, self.story.synopsis)
         self.assertRedirects(response, reverse("game", args=[game.id]))