+++ master/admin.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+
+import master.models as mmodels
+
+
+class StoryAdmin(admin.ModelAdmin):
+    fields = ["title", "slug", "synopsis", "main_conflict", "objective"]
+    list_display = ["title", "slug"]
+    prepopulated_fields = {"slug": ("title",)}
+
+
+admin.site.register(mmodels.Story, StoryAdmin)