+++ role_play/settings.py
@@ -34,6 +34,7 @@ ALLOWED_HOSTS = []
 INSTALLED_APPS = [
     "daphne",
     "game.apps.GameConfig",
+    "master.apps.MasterConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",