+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+### Added
+* The game master can create background stories before starting a game
+
 ## v0.3.0 - 2023-07-07
 
 ### Added